Faster build-times, smaller binary.
A lot of external dependencies have been disabled. The caveat is that this includes BLAS and LAPACK, which might impact the performance of pahmm if the following routines can be accelerated with these dependencies: dlib::find_min_box_constrained, dlib::bfgs_search_strategy, dlib::objective_delta_stop_strategy, dlib::find_min_bobyqa.
The pahmm Python library no longer uses a Debug-version of libpahmm, but rather a smaller and faster Release-version.
Build-times are faster due to the disabling of several dlib modules and the usage of the -j[cpu_count] flag with make.